+++ src/collections/components/CollectionProducts/CollectionProducts.tsx
@@ -35,9 +35,10 @@ const styles = (theme: Theme) =>
       "&:last-child": {
         paddingRight: 0
       },
-      width: 48 + theme.spacing.unit / 2
+      width: 76 + theme.spacing.unit / 2
     },
     colName: {
+      paddingLeft: 0,
       width: "auto"
     },
     colNameLabel: {